rework(layouts_dir): restructured layouts directory - decomposed it to patrial, singleton, subapp layouts; changed documentation for layouts directory;
