+++ README.md
@@ -7,18 +7,6 @@
 		- outActionName
 			Значит, что значение полученное из сервиса будет возвращаться вызывающей функции
 
-
-	- Принцип именования файлов в .../router/helpers/:
-		- Если файл экспортирует guard, то название генерируются следующим паттерном:
-			guardName.guard
-
-		- Если файл экспортирует сервисы для guard, то название генерируется следующим паттерном:
-			guardName + ExportFunctionalName = guardNameExportFunctionalName
-
-		- Пример:
-			auth.guard.js
-			authUpdateTokens.js
-
 	- Использование браузерных storage(sessionStorage, localStorage):
 		В случае необходимости использования какого-то ключа для storage используйте один из ключей объект
 		STORAGE_NAMES, который является полем consts; В случае если необходимого ключа там нет, создайте новую пару
@@ -95,12 +83,10 @@
 		- validators/
 
 	- layouts/
-		- DialogLayout/
-			- partial/ - дирректория в которой содержатся все диалоги
-			DialogLayout.vue
-		- EmptyLayout/ - layout без элементов
-		- MainLayout/ - главный layout
-		- ...
+		- partials/ - составные слои
+		- singletons/ - слои-синглтоны
+			- DialogLayout/
+				- partial/ - дирректория в подсистемах
 
 	- router/
 		- helpers/ - вспомогательный функционал для роутера
@@ -128,8 +114,6 @@
 <br>
 
 1. Если Issue успешно закрывается: <br>
-	Перед закрытием указывайте коммит, в котором она была решена, можно использовать следующий паттерн: <br>
-	Поправил в коммите 986ee1a
 
 2. Если Issue закрывается не успешно: <br>
 	Повесьте лэйбл wontfix и объясните причину закрытия issue
@@ -159,7 +143,7 @@
 
 <br>
 
-1. При коммите используйте следующий паттерн "task-type(область работы): сообщение" <br>
+1. При коммите используйте следующий паттерн "task-type(область работы): содержание" <br>
 
 <br><br>
