+++ src/layouts/singletons/DialogLayout/partial/TheAddToCartDialog/partial/FeatureSection.vue
