+++ src/layouts/singletons/DialogLayout/partial/TheAuthDialog/partial/SignInForm/SignInForm.vue
