+++ src/layouts/singletons/DialogLayout/partial/TheVerificationDialog.vue
