+++ src/layouts/subapps/HomeLayout/partial/TheMobileTape.vue
