+++ src/layouts/subapps/MainLayout/partial/TheStaticSidebar/TheStaticSidebar.vue
