+++ src/layouts/subapps/MainLayout/partial/TheStaticSidebar/partial/UserNavigation.vue
