+++ src/layouts/subapps/MainLayout/partial/TheStaticSidebarMobile/TheStaticSidebarMobile.vue
