+++ src/layouts/subapps/MainLayout/partial/TheStaticSidebarMobile/partial/CentralButtonIcons.vue
